+++ src/templates/user_settings_modal.js
